requirements.txt
langchain>=0.3.30
langchain-core>=1.1.0
langchain-community>=0.3.30
langchain-openai>=0.3.24
langchain-groq==1.1.1
langchain-pinecone>=0.2.8
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
-e .
